chart enter buy order decimal

+++ wp-content/themes/arbitrage-child/template-chart.php
@@ -2652,7 +2652,7 @@
 		
 		if (parseFloat(dbuypower) < (parseFloat(dpurprice) * parseFloat(dpurqty))) {
 			e.preventDefault();
-			$(".derrormes").text('You can only purchase a maximum of '+parseInt(dbuypower / dpurprice)+' stocks if the price is ₱'+dpurprice);
+			$(".derrormes").text('You can only purchase a maximum of '+parseFloat(dbuypower / dpurprice)+' stocks if the price is ₱'+dpurprice);
 			// console.log('You can only purchase a maximum of '+parseInt(dbuypower / dpurprice)+' stocks if the price is 	₱'+dpurprice);
 		}else {
 			if(x == 1 && y == 1){
